requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
